added slash commands and some changes

+++ cogs/general_commands.py
@@ -3,6 +3,7 @@ from discord.ext import commands
 import requests
 import re
 import json
+import asyncio
 
 class General_commands(commands.Cog):
   def __init__(self,client):
@@ -11,16 +12,24 @@ class General_commands(commands.Cog):
   @commands.Cog.listener()
   @commands.has_permissions(manage_roles=True, manage_messages=True)
   async def on_message(self, message):
-      role = message.guild.get_role(595403641058492416)
-      if message.author.bot:
-          pass  
-      elif role not in message.author.roles:
-          if re.search("knowledge is power",message.content,re.IGNORECASE):          
-              await message.author.add_roles(role)
-              embed = discord.Embed(title="Welcome Reader :book:",description=f"{message.author.mention} you now have the reader role. Feel free to explore the server.")
-              await message.channel.send(embed=embed)
-              await message.delete()
-  
+      if message.guild is None:
+        if message.author != self.client.user:
+          pass
+
+
+      else:
+          if message.guild.id == 584581061271617538:
+              if message.author.bot:
+                  pass  
+              else:
+                  role = message.guild.get_role(595403641058492416)
+                  if role not in message.author.roles:
+                      if re.search("knowledge is power",message.content,re.IGNORECASE):          
+                          await message.author.add_roles(role)
+                          embed = discord.Embed(title="Welcome Reader :book:",description=f"{message.author.mention} you now have the reader role. Feel free to explore the server.")
+                          await message.channel.send(embed=embed)
+                          await message.delete()
+          
   @commands.command(brief='Sends an inspirational quote',description='Sends an inspirational quote')
   async def inspire(self, ctx):
       try:
@@ -39,16 +48,14 @@ class General_commands(commands.Cog):
           embed = discord.Embed(title="Here to inspire 🌈", description= quote)
           await ctx.send(embed=embed)
   
-  # @commands.command(brief='Assigns the Theoretical Reader role',description='Assigns the reader role to people who get the role right')
-  # async def code(self,ctx):
-  #     if re.search("knowledge is power",ctx.message.content,re.IGNORECASE) :
-  #         role = ctx.guild.get_role(595403641058492416)
-  #         await ctx.author.add_roles(role)
-  #         embed = discord.Embed(title="Welcome Reader :book:",description=f"Gave the role {role.name} to {ctx.author.mention}")
-  #         await ctx.send(embed=embed)
-  #         await ctx.message.delete()
-  #     else:
-  #         await ctx.send("Hmmm that's not right. Here's a hint: it's a 3 part code.")
 
+  @commands.command(brief="ping the bot", description="ping the bot")
+  async def ping(self,ctx):
+      await ctx.send("Pong!")
+
+
+  @commands.command(brief="pong the bot", description="pong the bot")
+  async def pong(self,ctx):
+      await ctx.send("Ping!")
 def setup(client):
   client.add_cog(General_commands(client))
